refactor(core): move inpaint + remaining CV modules to nukebg-core (phase 7)
Batch C + remainder: move inpaint-telea/blend, patchmatch-inpaint, simple/grid-flood-fill, alpha-matting/refine, foreground-estimation, shadow-cleanup, signature-threshold, subject-exclusion, lama-crop/router (+ tests) into core/src/cv. Extract pure patchMatchInpaint into core/src/inpaint/patch-match.ts; inpaint.worker.ts keeps a thin shell importing from core. Rewire 6 app importers + 5 app tests. Milestone: all 21 CV modules + patchmatch in core.

KNOWN DEBT: nukebg-core standalone build (tsc -b) currently reports ~420 noUncheckedIndexedAccess errors in the moved legacy CV code (written under the app's looser tsconfig). CI gate is green (app typecheck/test/lint), but core cannot build standalone yet — must be resolved before Phase 19 (distribution): either relax noUncheckedIndexedAccess for core or add assertions.

Closes #272
