+++ packages/nukebg-app/src/pipeline/orchestrator.ts
@@ -16,8 +16,8 @@ import type {
 } from '../types/worker-messages';
 import { IMAGE_CLASSIFY_PARAMS, INPAINT_PARAMS, PRECISION_PROFILES } from 'nukebg-core';
 import type { PrecisionMode } from 'nukebg-core';
-import { compositeWithFeather, dilateMask } from '../workers/cv/inpaint-blend';
-import { shouldUseLama } from '../workers/cv/lama-router';
+import { compositeWithFeather, dilateMask } from 'nukebg-core/cv/inpaint-blend';
+import { shouldUseLama } from 'nukebg-core/cv/lama-router';
 import { WorkerChannel } from './worker-channel';
 import type { ImageProcessor, StageCallback } from './image-processor';
