+++ packages/nukebg-app/src/workers/cv.worker.ts
@@ -2,10 +2,10 @@ import { detectBgColors } from 'nukebg-core/cv/detect-bg-colors';
 import { watermarkDetect } from 'nukebg-core/cv/watermark-detect';
 import { watermarkDetectDalle } from 'nukebg-core/cv/watermark-dalle';
 import { sparkleDetect } from 'nukebg-core/cv/sparkle-detect';
-import { alphaRefine } from './cv/alpha-refine';
+import { alphaRefine } from 'nukebg-core/cv/alpha-refine';
 import { extractImageFeatures, classifyImage } from 'nukebg-core/cv/classify-image';
-import { signatureThreshold } from './cv/signature-threshold';
-import { estimateForeground } from './cv/foreground-estimation';
+import { signatureThreshold } from 'nukebg-core/cv/signature-threshold';
+import { estimateForeground } from 'nukebg-core/cv/foreground-estimation';
 import type { CvWorkerRequest } from '../types/worker-messages';
 
 self.onmessage = (e: MessageEvent<CvWorkerRequest>) => {
