+++ packages/nukebg-core/src/cv/signature-threshold.ts
@@ -1,4 +1,4 @@
-import { IMAGE_CLASSIFY_PARAMS } from 'nukebg-core';
+import { IMAGE_CLASSIFY_PARAMS } from '../pipeline/constants.js';
 
 /**
  * Signature threshold: extract signature strokes from a mostly-white background
